refactor: split preferences module into mod.rs + tests.rs
